refactor: extract KeychainSecretStoring.swift from KeychainSecretStore.swift
Splits the `KeychainSecretStoring` protocol (with the load-bearing
`SettingsStore` seam doc-comment) out of KeychainSecretStore.swift so
that file holds only the LegacyDeletionFailure data value + the concrete
KeychainSecretStore impl.

Byte-identical move of lines 11-42. Both files import Foundation.

KeychainSecretStore.swift: 103 → 71 lines. Tests: 667.
